test: 테스트 디렉토리 개편
Refs: CR-80
