+++ tests/context_extractor/csharp/SampleCalculator.cs
