+++ tests/context_extractor/csharp/test_csharp_context_extraction.py
@@ -1,4 +1,4 @@
-"""ContextExtractor Go 테스트 케이스."""
+"""ContextExtractor C# 테스트 케이스."""
 
 from __future__ import annotations
 
@@ -15,31 +15,34 @@ class TestBasicFunctionExtraction:
     @pytest.fixture
     def sample_file_path(self) -> Path:
         """테스트용 샘플 파일 경로를 반환합니다."""
-        return Path(__file__).parent / "language_samples" / "go" / "SampleCalculator.go"
+        return (
+            Path(__file__).parent
+            / "SampleCalculator.cs"
+        )
 
     @pytest.fixture
     def extractor(self) -> ContextExtractor:
-        """Go용 ContextExtractor 인스턴스를 반환합니다."""
-        return ContextExtractor("go")
+        """C#용 ContextExtractor 인스턴스를 반환합니다."""
+        return ContextExtractor("csharp")
 
     def test_class_declaration(
         self,
         extractor: ContextExtractor,
         sample_file_path: Path,
     ) -> None:
-        """구조체 선언부 추출 테스트."""
-        changed_ranges = [LineRange(33, 33)]  # SampleCalculator 구조체 선언부
+        """클래스 선언부 추출 테스트."""
+        changed_ranges = [LineRange(32, 32)]  # SampleCalculator 클래스 선언부
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        # import 문 검증
-        assert "fmt" in all_context
-        assert "math" in all_context
-        assert "strings" in all_context
+        # using 문 검증
+        assert "using System;" in all_context
+        assert "using System.Collections.Generic;" in all_context
+        assert "using System.Linq;" in all_context
         assert (
-            "type SampleCalculator struct" in all_context
-            or "SampleCalculator struct" in all_context
+            "public class SampleCalculator" in all_context
+            or "class SampleCalculator" in all_context
         )
 
     def test_constructor_method(
@@ -47,19 +50,22 @@ class TestBasicFunctionExtraction:
         extractor: ContextExtractor,
         sample_file_path: Path,
     ) -> None:
-        """New 함수 추출 테스트."""
-        changed_ranges = [LineRange(42, 51)]  # NewSampleCalculator 함수
+        """생성자 메서드 추출 테스트."""
+        changed_ranges = [LineRange(46, 54)]  # 생성자
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        # import 문 검증
-        assert "fmt" in all_context
-        assert "math" in all_context
-        assert "strings" in all_context
-        # 구조체 선언부 검증
-        assert "type SampleCalculator struct" in all_context
-        assert "func NewSampleCalculator" in all_context or "func" in all_context
+        # using 문 검증
+        assert "using System;" in all_context
+        assert "using System.Collections.Generic;" in all_context
+        assert "using System.Linq;" in all_context
+        # 부모 클래스 선언부 검증
+        assert "public class SampleCalculator" in all_context
+        assert (
+            "public SampleCalculator" in all_context
+            or "SampleCalculator(" in all_context
+        )
         assert "value" in all_context
         assert "history" in all_context
 
@@ -68,19 +74,20 @@ class TestBasicFunctionExtraction:
         extractor: ContextExtractor,
         sample_file_path: Path,
     ) -> None:
-        """메서드(리시버 함수) 추출 테스트."""
-        changed_ranges = [LineRange(53, 82)]  # AddNumbers 메서드
+        """클래스 메서드 추출 테스트."""
+        changed_ranges = [LineRange(56, 90)]  # AddNumbers 메서드
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        # import 문 검증
-        assert "fmt" in all_context
-        assert "math" in all_context
-        assert "strings" in all_context
-        # 구조체 선언부 검증
-        assert "type SampleCalculator struct" in all_context
-        assert "func" in all_context and "AddNumbers" in all_context
+        # using 문 검증
+        assert "using System;" in all_context
+        assert "using System.Collections.Generic;" in all_context
+        assert "using System.Linq;" in all_context
+        # 부모 클래스 선언부 검증
+        assert "public class SampleCalculator" in all_context
+        assert "AddNumbers" in all_context
+        assert "public" in all_context or "private" in all_context
 
     def test_complex_method(
         self,
@@ -88,13 +95,17 @@ class TestBasicFunctionExtraction:
         sample_file_path: Path,
     ) -> None:
         """복잡한 메서드 추출 테스트."""
-        changed_ranges = [LineRange(84, 133)]  # MultiplyAndFormat 메서드
+        changed_ranges = [LineRange(92, 149)]  # MultiplyAndFormat 메서드
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        # 구조체 선언부 검증
-        assert "type SampleCalculator struct" in all_context
+        # using 문 검증
+        assert "using System;" in all_context
+        assert "using System.Collections.Generic;" in all_context
+        assert "using System.Linq;" in all_context
+        # 부모 클래스 선언부 검증
+        assert "public class SampleCalculator" in all_context
         assert "MultiplyAndFormat" in all_context
 
     def test_nested_inner_function(
@@ -103,37 +114,37 @@ class TestBasicFunctionExtraction:
         sample_file_path: Path,
     ) -> None:
         """중첩 내부 함수 추출 테스트."""
-        changed_ranges = [LineRange(96, 102)]  # multiplyRecursive 내부 함수
+        changed_ranges = [LineRange(107, 114)]  # MultiplyRecursive 내부 함수
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        # 구조체 선언부 검증
-        assert "type SampleCalculator struct" in all_context
+        # 부모 클래스 선언부 검증
+        assert "public class SampleCalculator" in all_context
         # 부모 메서드 선언부 검증
         assert "MultiplyAndFormat" in all_context
-        assert "multiplyRecursive" in all_context
+        assert "MultiplyRecursive" in all_context
 
     def test_external_function(
         self,
         extractor: ContextExtractor,
         sample_file_path: Path,
     ) -> None:
-        """외부 함수 추출 테스트."""
-        changed_ranges = [LineRange(154, 170)]  # HelperFunction
+        """정적 메서드 추출 테스트."""
+        changed_ranges = [LineRange(179, 195)]  # HelperFunction
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        assert "func HelperFunction" in all_context or "HelperFunction" in all_context
+        assert "static" in all_context and "HelperFunction" in all_context
 
     def test_factory_function(
         self,
         extractor: ContextExtractor,
         sample_file_path: Path,
     ) -> None:
-        """팩토리 함수 추출 테스트."""
-        changed_ranges = [LineRange(172, 201)]  # AdvancedCalculatorFactory
+        """팩토리 메서드 추출 테스트."""
+        changed_ranges = [LineRange(204, 230)]  # AdvancedCalculatorFactory
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
@@ -146,13 +157,13 @@ class TestBasicFunctionExtraction:
         sample_file_path: Path,
     ) -> None:
         """메서드 선언부만 추출 테스트."""
-        changed_ranges = [LineRange(53, 53)]  # AddNumbers 선언부만
+        changed_ranges = [LineRange(56, 56)]  # AddNumbers 선언부만
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        # 구조체 선언부 검증
-        assert "type SampleCalculator struct" in all_context
+        # 부모 클래스 선언부 검증
+        assert "public class SampleCalculator" in all_context
         assert "AddNumbers" in all_context
 
     def test_external_function_declaration_only(
@@ -160,8 +171,8 @@ class TestBasicFunctionExtraction:
         extractor: ContextExtractor,
         sample_file_path: Path,
     ) -> None:
-        """외부 함수 선언부만 추출 테스트."""
-        changed_ranges = [LineRange(154, 154)]  # HelperFunction 선언부만
+        """정적 메서드 선언부만 추출 테스트."""
+        changed_ranges = [LineRange(179, 179)]  # HelperFunction 선언부만
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
@@ -175,12 +186,15 @@ class TestModuleLevelElements:
     @pytest.fixture
     def sample_file_path(self) -> Path:
         """테스트용 샘플 파일 경로를 반환합니다."""
-        return Path(__file__).parent / "language_samples" / "go" / "SampleCalculator.go"
+        return (
+            Path(__file__).parent
+            / "SampleCalculator.cs"
+        )
 
     @pytest.fixture
     def extractor(self) -> ContextExtractor:
-        """Go용 ContextExtractor 인스턴스를 반환합니다."""
-        return ContextExtractor("go")
+        """C#용 ContextExtractor 인스턴스를 반환합니다."""
+        return ContextExtractor("csharp")
 
     def test_basic_constants(
         self,
@@ -188,26 +202,26 @@ class TestModuleLevelElements:
         sample_file_path: Path,
     ) -> None:
         """기본 상수들 추출 테스트."""
-        changed_ranges = [LineRange(14, 18)]  # 상수 선언
+        changed_ranges = [LineRange(10, 22)]  # Constants 클래스
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        assert "MaxCalculationSteps" in all_context or "const" in all_context
-        assert "DefaultPrecision" in all_context
+        assert "MAX_CALCULATION_STEPS" in all_context or "const" in all_context
+        assert "DEFAULT_PRECISION" in all_context
 
     def test_object_constant(
         self,
         extractor: ContextExtractor,
         sample_file_path: Path,
     ) -> None:
-        """구조체 상수 추출 테스트."""
-        changed_ranges = [LineRange(20, 24)]  # CalculationModes 구조체
+        """객체 상수 추출 테스트."""
+        changed_ranges = [LineRange(16, 21)]  # CALCULATION_MODES 객체
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        assert "CalculationModes" in all_context
+        assert "CALCULATION_MODES" in all_context
 
     def test_module_bottom_constants(
         self,
@@ -215,12 +229,12 @@ class TestModuleLevelElements:
         sample_file_path: Path,
     ) -> None:
         """모듈 하단 상수들 추출 테스트."""
-        changed_ranges = [LineRange(204, 204)]  # ModuleVersion
+        changed_ranges = [LineRange(234, 242)]  # ModuleConstants
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        assert "ModuleVersion" in all_context
+        assert "MODULE_VERSION" in all_context
 
 
 class TestMultiRangeExtraction:
@@ -229,21 +243,24 @@ class TestMultiRangeExtraction:
     @pytest.fixture
     def sample_file_path(self) -> Path:
         """테스트용 샘플 파일 경로를 반환합니다."""
-        return Path(__file__).parent / "language_samples" / "go" / "SampleCalculator.go"
+        return (
+            Path(__file__).parent
+            / "SampleCalculator.cs"
+        )
 
     @pytest.fixture
     def extractor(self) -> ContextExtractor:
-        """Go용 ContextExtractor 인스턴스를 반환합니다."""
-        return ContextExtractor("go")
+        """C#용 ContextExtractor 인스턴스를 반환합니다."""
+        return ContextExtractor("csharp")
 
     def test_three_cross_functions(
         self,
         extractor: ContextExtractor,
         sample_file_path: Path,
     ) -> None:
-        """3개 함수에 걸친 영역 추출 테스트."""
+        """3개 메서드에 걸친 영역 추출 테스트."""
         changed_ranges = [
-            LineRange(135, 201)
+            LineRange(151, 230)
         ]  # CalculateCircleArea ~ AdvancedCalculatorFactory
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
@@ -259,7 +276,7 @@ class TestMultiRangeExtraction:
         sample_file_path: Path,
     ) -> None:
         """2개 블록에 걸친 메서드 추출 테스트."""
-        changed_ranges = [LineRange(53, 61), LineRange(154, 158)]
+        changed_ranges = [LineRange(56, 64), LineRange(179, 185)]
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
@@ -274,16 +291,16 @@ class TestMultiRangeExtraction:
     ) -> None:
         """비연속적인 여러 범위 추출 테스트."""
         changed_ranges = [
-            LineRange(14, 18),  # 파일 상수들
-            LineRange(141, 143),  # validateRadius 내부 함수
-            LineRange(204, 204),  # 모듈 레벨 상수들
+            LineRange(12, 14),  # 파일 상수들
+            LineRange(158, 161),  # ValidateRadius 내부 함수
+            LineRange(236, 238),  # 모듈 레벨 상수들
         ]
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        assert "MaxCalculationSteps" in all_context or "const" in all_context
-        assert "ModuleVersion" in all_context
+        assert "MAX_CALCULATION_STEPS" in all_context or "const" in all_context
+        assert "MODULE_VERSION" in all_context
 
 
 class TestComplexScenarios:
@@ -292,26 +309,29 @@ class TestComplexScenarios:
     @pytest.fixture
     def sample_file_path(self) -> Path:
         """테스트용 샘플 파일 경로를 반환합니다."""
-        return Path(__file__).parent / "language_samples" / "go" / "SampleCalculator.go"
+        return (
+            Path(__file__).parent
+            / "SampleCalculator.cs"
+        )
 
     @pytest.fixture
     def extractor(self) -> ContextExtractor:
-        """Go용 ContextExtractor 인스턴스를 반환합니다."""
-        return ContextExtractor("go")
+        """C#용 ContextExtractor 인스턴스를 반환합니다."""
+        return ContextExtractor("csharp")
 
     def test_entire_class_extraction(
         self,
         extractor: ContextExtractor,
         sample_file_path: Path,
     ) -> None:
-        """전체 구조체와 관련 메서드 추출 테스트."""
-        changed_ranges = [LineRange(33, 133)]  # SampleCalculator 전체 구조체와 메서드들
+        """전체 클래스 추출 테스트."""
+        changed_ranges = [LineRange(32, 171)]  # SampleCalculator 전체 클래스
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        assert "SampleCalculator" in all_context
-        assert "func" in all_context
+        assert "class SampleCalculator" in all_context
+        assert "SampleCalculator(" in all_context
         assert "AddNumbers" in all_context
 
     def test_class_and_module_constants(
@@ -319,14 +339,14 @@ class TestComplexScenarios:
         extractor: ContextExtractor,
         sample_file_path: Path,
     ) -> None:
-        """구조체와 모듈 상수 동시 추출 테스트."""
-        changed_ranges = [LineRange(14, 209)]  # 상수부터 모듈 끝까지
+        """클래스와 모듈 상수 동시 추출 테스트."""
+        changed_ranges = [LineRange(10, 242)]  # 상수부터 모듈 끝까지
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
         assert "SampleCalculator" in all_context
-        assert "ModuleVersion" in all_context
+        assert "MODULE_VERSION" in all_context
 
 
 class TestEdgeCases:
@@ -335,12 +355,15 @@ class TestEdgeCases:
     @pytest.fixture
     def sample_file_path(self) -> Path:
         """테스트용 샘플 파일 경로를 반환합니다."""
-        return Path(__file__).parent / "language_samples" / "go" / "SampleCalculator.go"
+        return (
+            Path(__file__).parent
+            / "SampleCalculator.cs"
+        )
 
     @pytest.fixture
     def extractor(self) -> ContextExtractor:
-        """Go용 ContextExtractor 인스턴스를 반환합니다."""
-        return ContextExtractor("go")
+        """C#용 ContextExtractor 인스턴스를 반환합니다."""
+        return ContextExtractor("csharp")
 
     def test_invalid_line_ranges(
         self,
