+++ tests/context_extractor/java/test_java_context_extraction.py
@@ -1,4 +1,4 @@
-"""ContextExtractor Swift 테스트 케이스."""
+"""ContextExtractor Java 테스트 케이스."""
 
 from __future__ import annotations
 
@@ -17,15 +17,13 @@ class TestBasicFunctionExtraction:
         """테스트용 샘플 파일 경로를 반환합니다."""
         return (
             Path(__file__).parent
-            / "language_samples"
-            / "swift"
-            / "SampleCalculator.swift"
+            / "SampleCalculator.java"
         )
 
     @pytest.fixture
     def extractor(self) -> ContextExtractor:
-        """Swift용 ContextExtractor 인스턴스를 반환합니다."""
-        return ContextExtractor("swift")
+        """Java용 ContextExtractor 인스턴스를 반환합니다."""
+        return ContextExtractor("java")
 
     def test_class_declaration(
         self,
@@ -33,32 +31,39 @@ class TestBasicFunctionExtraction:
         sample_file_path: Path,
     ) -> None:
         """클래스 선언부 추출 테스트."""
-        changed_ranges = [LineRange(25, 25)]  # SampleCalculator 클래스 선언부
+        changed_ranges = [LineRange(20, 20)]  # SampleCalculator 클래스 선언부
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
         # import 문 검증
-        assert "import Foundation" in all_context
-        assert "class SampleCalculator" in all_context
+        assert "import java.util.*;" in all_context
+        # 클래스 검증 (Java는 전체 클래스가 추출됨)
+        assert "public class SampleCalculator" in all_context
+        # Java에서는 클래스 전체가 추출되므로 내부 요소들도 포함됨
+        assert "private int value" in all_context
 
     def test_constructor_method(
         self,
         extractor: ContextExtractor,
         sample_file_path: Path,
     ) -> None:
-        """init 메서드 추출 테스트."""
-        changed_ranges = [LineRange(34, 41)]  # init 메서드
+        """생성자 메서드 추출 테스트."""
+        changed_ranges = [LineRange(33, 40)]  # 파라미터 있는 생성자
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
         # import 문 검증
-        assert "import Foundation" in all_context
-        assert "class SampleCalculator" in all_context
-        assert "init(" in all_context
-        assert "value" in all_context
-        assert "history" in all_context
+        assert "import java.util.*;" in all_context
+        # 부모 클래스 선언부 검증
+        assert "public class SampleCalculator" in all_context
+        # 선언부 검증
+        assert "public SampleCalculator(int initialValue)" in all_context
+        # 내부 코드 블록 검증 (전체 메서드 추출 확인)
+        assert "this.value = initialValue" in all_context
+        assert "this.history = new ArrayList<>()" in all_context
+        assert 'this.mode = Constants.CALCULATION_MODES.get("basic")' in all_context
 
     def test_class_method(
         self,
@@ -66,15 +71,21 @@ class TestBasicFunctionExtraction:
         sample_file_path: Path,
     ) -> None:
         """클래스 메서드 추출 테스트."""
-        changed_ranges = [LineRange(43, 72)]  # addNumbers 메서드
+        changed_ranges = [LineRange(42, 44)]  # addNumbers 메서드
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
         # import 문 검증
-        assert "import Foundation" in all_context
-        assert "class SampleCalculator" in all_context
-        assert "func addNumbers" in all_context
+        assert "import java.util.*;" in all_context
+        # 부모 클래스 선언부 검증
+        assert "public class SampleCalculator" in all_context
+        # 선언부 검증
+        assert "public int addNumbers(int a, int b)" in all_context
+        # 내부 코드 블록 검증 (전체 메서드 추출 확인)
+        assert "class InputValidator" in all_context
+        assert "class OperationLogger" in all_context
+        assert "int result = a + b" in all_context
 
     def test_complex_method(
         self,
@@ -82,13 +93,27 @@ class TestBasicFunctionExtraction:
         sample_file_path: Path,
     ) -> None:
         """복잡한 메서드 추출 테스트."""
-        changed_ranges = [LineRange(74, 126)]  # multiplyAndFormat 메서드
+        changed_ranges = [
+            LineRange(75, 78),
+            LineRange(120, 125),
+        ]  # multiplyAndFormat 메서드
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        assert "class SampleCalculator" in all_context
-        assert "func multiplyAndFormat" in all_context
+        # import 문 검증
+        assert "import java.util.*;" in all_context
+        # 부모 클래스 선언부 검증
+        assert "public class SampleCalculator" in all_context
+        # 선언부 검증
+        assert (
+            "public Map<String, Object> multiplyAndFormat(List<Integer> numbers)"
+            in all_context
+        )
+        # 내부 코드 블록 검증 (전체 메서드 추출 확인)
+        assert "class ProductCalculator" in all_context
+        assert "class ResultFormatter" in all_context
+        assert "class RecursiveMultiplier" in all_context
 
     def test_nested_inner_function(
         self,
@@ -96,31 +121,47 @@ class TestBasicFunctionExtraction:
         sample_file_path: Path,
     ) -> None:
         """중첩 내부 함수 추출 테스트."""
-        changed_ranges = [LineRange(86, 91)]  # multiplyRecursive 내부 함수
+        changed_ranges = [LineRange(87, 95)]  # RecursiveMultiplier 내부 클래스
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
+        # import 문 검증
+        assert "import java.util.*;" in all_context
         # 부모 클래스 선언부 검증
-        assert "class SampleCalculator" in all_context
-        assert "func multiplyAndFormat" in all_context
+        assert "public class SampleCalculator" in all_context
+        # 부모 메서드 선언부 검증
         assert (
-            "func multiplyRecursive" in all_context
-            or "multiplyRecursive" in all_context
+            "public Map<String, Object> multiplyAndFormat(List<Integer> numbers)"
+            in all_context
         )
+        # 선언부 검증
+        assert (
+            "class ProductCalculator" in all_context
+            or "class RecursiveMultiplier" in all_context
+        )
+        # 내부 코드 블록 검증 (전체 함수 추출 확인)
+        assert "public int multiplyRecursive" in all_context
+        assert "if (index >= items.size())" in all_context
 
     def test_external_function(
         self,
         extractor: ContextExtractor,
         sample_file_path: Path,
     ) -> None:
-        """외부 함수 추출 테스트."""
-        changed_ranges = [LineRange(148, 164)]  # helperFunction
+        """클래스 외부 함수 추출 테스트."""
+        changed_ranges = [LineRange(155, 160)]  # helperFunction 내부 코드 범위
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        assert "func helperFunction" in all_context
+        # import 문 검증
+        assert "import java.util.*;" in all_context
+        # 선언부 검증
+        assert "public static String helperFunction" in all_context
+        # 내부 코드 블록 검증 (전체 함수 추출 확인)
+        assert "class DictFormatter" in all_context
+        assert "List<String> formattedItems" in all_context
 
     def test_factory_function(
         self,
@@ -128,12 +169,19 @@ class TestBasicFunctionExtraction:
         sample_file_path: Path,
     ) -> None:
         """팩토리 함수 추출 테스트."""
-        changed_ranges = [LineRange(166, 188)]  # advancedCalculatorFactory
+        changed_ranges = [LineRange(182, 190)]  # advancedCalculatorFactory 메서드
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        assert "func advancedCalculatorFactory" in all_context
+        # 선언부 검증
+        assert (
+            "public static SampleCalculator advancedCalculatorFactory(String mode)"
+            in all_context
+        )
+        # 내부 코드 블록 검증 (전체 함수 추출 확인)
+        assert "class CalculatorCreator" in all_context
+        assert "class ModeValidator" in all_context
 
     def test_method_declaration_only(
         self,
@@ -141,13 +189,16 @@ class TestBasicFunctionExtraction:
         sample_file_path: Path,
     ) -> None:
         """메서드 선언부만 추출 테스트."""
-        changed_ranges = [LineRange(43, 43)]  # addNumbers 선언부만
+        changed_ranges = [LineRange(42, 42)]  # addNumbers 선언부만
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        assert "class SampleCalculator" in all_context
-        assert "addNumbers" in all_context
+        # 부모 클래스 선언부 검증
+        assert "public class SampleCalculator" in all_context
+        assert "public int addNumbers(int a, int b)" in all_context
+        # Java에서는 메서드 전체가 추출되므로 내부 클래스도 포함됨
+        assert "class InputValidator" in all_context
 
     def test_external_function_declaration_only(
         self,
@@ -155,12 +206,14 @@ class TestBasicFunctionExtraction:
         sample_file_path: Path,
     ) -> None:
         """외부 함수 선언부만 추출 테스트."""
-        changed_ranges = [LineRange(148, 148)]  # helperFunction 선언부만
+        changed_ranges = [LineRange(155, 155)]  # helperFunction 선언부만
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        assert "helperFunction" in all_context
+        assert "public static String helperFunction" in all_context
+        # Java에서는 메서드 전체가 추출되므로 내부 클래스도 포함됨
+        assert "class DictFormatter" in all_context
 
 
 class TestModuleLevelElements:
@@ -171,15 +224,13 @@ class TestModuleLevelElements:
         """테스트용 샘플 파일 경로를 반환합니다."""
         return (
             Path(__file__).parent
-            / "language_samples"
-            / "swift"
-            / "SampleCalculator.swift"
+            / "SampleCalculator.java"
         )
 
     @pytest.fixture
     def extractor(self) -> ContextExtractor:
-        """Swift용 ContextExtractor 인스턴스를 반환합니다."""
-        return ContextExtractor("swift")
+        """Java용 ContextExtractor 인스턴스를 반환합니다."""
+        return ContextExtractor("java")
 
     def test_basic_constants(
         self,
@@ -187,26 +238,26 @@ class TestModuleLevelElements:
         sample_file_path: Path,
     ) -> None:
         """기본 상수들 추출 테스트."""
-        changed_ranges = [LineRange(8, 10)]  # 상수 선언
+        changed_ranges = [LineRange(9, 10)]  # Constants 클래스 내 상수들
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        assert "let" in all_context
-        assert "100" in all_context or "2" in all_context
+        assert "MAX_CALCULATION_STEPS = 100" in all_context
+        assert "DEFAULT_PRECISION = 2" in all_context
 
-    def test_object_constant(
+    def test_map_constant(
         self,
         extractor: ContextExtractor,
         sample_file_path: Path,
     ) -> None:
-        """구조체 상수 추출 테스트."""
-        changed_ranges = [LineRange(12, 16)]  # CALCULATION_MODES 배열
+        """맵 상수 추출 테스트."""
+        changed_ranges = [LineRange(12, 16)]  # CALCULATION_MODES 맵
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        assert "CALCULATION_MODES" in all_context or "let" in all_context
+        assert "CALCULATION_MODES = Map.of(" in all_context
 
     def test_module_bottom_constants(
         self,
@@ -214,12 +265,12 @@ class TestModuleLevelElements:
         sample_file_path: Path,
     ) -> None:
         """모듈 하단 상수들 추출 테스트."""
-        changed_ranges = [LineRange(191, 195)]  # MODULE_VERSION
+        changed_ranges = [LineRange(212, 213)]  # MODULE_VERSION
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        assert "MODULE_VERSION" in all_context
+        assert 'MODULE_VERSION = "1.0.0"' in all_context
 
 
 class TestMultiRangeExtraction:
@@ -230,32 +281,30 @@ class TestMultiRangeExtraction:
         """테스트용 샘플 파일 경로를 반환합니다."""
         return (
             Path(__file__).parent
-            / "language_samples"
-            / "swift"
-            / "SampleCalculator.swift"
+            / "SampleCalculator.java"
         )
 
     @pytest.fixture
     def extractor(self) -> ContextExtractor:
-        """Swift용 ContextExtractor 인스턴스를 반환합니다."""
-        return ContextExtractor("swift")
+        """Java용 ContextExtractor 인스턴스를 반환합니다."""
+        return ContextExtractor("java")
 
-    def test_three_cross_functions(
+    def test_three_cross_classes(
         self,
         extractor: ContextExtractor,
         sample_file_path: Path,
     ) -> None:
-        """3개 함수에 걸친 영역 추출 테스트."""
+        """3개 클래스에 걸친 영역 추출 테스트."""
         changed_ranges = [
-            LineRange(128, 188)
-        ]  # calculateCircleArea ~ advancedCalculatorFactory
+            LineRange(129, 210)
+        ]  # calculateCircleArea ~ CalculatorFactory
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        assert "calculateCircleArea" in all_context
-        assert "helperFunction" in all_context
-        assert "advancedCalculatorFactory" in all_context
+        assert "public double calculateCircleArea" in all_context
+        assert "public static String helperFunction" in all_context
+        assert "public static SampleCalculator advancedCalculatorFactory" in all_context
 
     def test_two_blocks_cross_methods(
         self,
@@ -263,13 +312,13 @@ class TestMultiRangeExtraction:
         sample_file_path: Path,
     ) -> None:
         """2개 블록에 걸친 메서드 추출 테스트."""
-        changed_ranges = [LineRange(43, 51), LineRange(148, 153)]
+        changed_ranges = [LineRange(42, 47), LineRange(155, 160)]
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 2
         all_context = "\n".join(contexts)
-        assert "addNumbers" in all_context
-        assert "helperFunction" in all_context
+        assert "public int addNumbers(int a, int b)" in all_context
+        assert "public static String helperFunction" in all_context
 
     def test_non_contiguous_ranges(
         self,
@@ -278,16 +327,17 @@ class TestMultiRangeExtraction:
     ) -> None:
         """비연속적인 여러 범위 추출 테스트."""
         changed_ranges = [
-            LineRange(8, 10),  # 파일 상수들
-            LineRange(134, 136),  # validateRadius 내부 함수
-            LineRange(191, 195),  # 모듈 레벨 상수들
+            LineRange(9, 10),  # Constants 클래스 상수들
+            LineRange(135, 137),  # RadiusValidator 내부 클래스
+            LineRange(212, 213),  # 모듈 레벨 상수들
         ]
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 2
         all_context = "\n".join(contexts)
-        assert "MAX_CALCULATION_STEPS" in all_context or "let" in all_context
-        assert "MODULE_VERSION" in all_context
+        assert "MAX_CALCULATION_STEPS = 100" in all_context
+        assert "class RadiusValidator" in all_context or "validateRadius" in all_context
+        assert 'MODULE_VERSION = "1.0.0"' in all_context
 
 
 class TestComplexScenarios:
@@ -298,15 +348,13 @@ class TestComplexScenarios:
         """테스트용 샘플 파일 경로를 반환합니다."""
         return (
             Path(__file__).parent
-            / "language_samples"
-            / "swift"
-            / "SampleCalculator.swift"
+            / "SampleCalculator.java"
         )
 
     @pytest.fixture
     def extractor(self) -> ContextExtractor:
-        """Swift용 ContextExtractor 인스턴스를 반환합니다."""
-        return ContextExtractor("swift")
+        """Java용 ContextExtractor 인스턴스를 반환합니다."""
+        return ContextExtractor("java")
 
     def test_entire_class_extraction(
         self,
@@ -314,14 +362,14 @@ class TestComplexScenarios:
         sample_file_path: Path,
     ) -> None:
         """전체 클래스 추출 테스트."""
-        changed_ranges = [LineRange(25, 146)]  # SampleCalculator 전체 클래스
+        changed_ranges = [LineRange(20, 148)]  # SampleCalculator 전체 클래스
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        assert "class SampleCalculator" in all_context
-        assert "init(" in all_context
-        assert "addNumbers" in all_context
+        assert "public class SampleCalculator" in all_context
+        assert "public SampleCalculator(int initialValue)" in all_context
+        assert "public int addNumbers(int a, int b)" in all_context
 
     def test_class_and_module_constants(
         self,
@@ -329,13 +377,13 @@ class TestComplexScenarios:
         sample_file_path: Path,
     ) -> None:
         """클래스와 모듈 상수 동시 추출 테스트."""
-        changed_ranges = [LineRange(8, 195)]  # 상수부터 모듈 끝까지
+        changed_ranges = [LineRange(8, 217)]  # Constants 클래스부터 모듈 상수까지
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        assert "SampleCalculator" in all_context
-        assert "MODULE_VERSION" in all_context
+        assert "public class SampleCalculator" in all_context
+        assert 'MODULE_VERSION = "1.0.0"' in all_context
 
 
 class TestEdgeCases:
@@ -346,15 +394,13 @@ class TestEdgeCases:
         """테스트용 샘플 파일 경로를 반환합니다."""
         return (
             Path(__file__).parent
-            / "language_samples"
-            / "swift"
-            / "SampleCalculator.swift"
+            / "SampleCalculator.java"
         )
 
     @pytest.fixture
     def extractor(self) -> ContextExtractor:
-        """Swift용 ContextExtractor 인스턴스를 반환합니다."""
-        return ContextExtractor("swift")
+        """Java용 ContextExtractor 인스턴스를 반환합니다."""
+        return ContextExtractor("java")
 
     def test_invalid_line_ranges(
         self,
