+++ tests/context_extractor/javascript/SampleCalculator.js
