+++ tests/context_extractor/kotlin/SampleCalculator.kt
