+++ tests/context_extractor/python/sample_class.py
