+++ tests/context_extractor/swift/test_swift_context_extraction.py
@@ -1,4 +1,4 @@
-"""ContextExtractor C# 테스트 케이스."""
+"""ContextExtractor Swift 테스트 케이스."""
 
 from __future__ import annotations
 
@@ -17,15 +17,13 @@ class TestBasicFunctionExtraction:
         """테스트용 샘플 파일 경로를 반환합니다."""
         return (
             Path(__file__).parent
-            / "language_samples"
-            / "csharp"
-            / "SampleCalculator.cs"
+            / "SampleCalculator.swift"
         )
 
     @pytest.fixture
     def extractor(self) -> ContextExtractor:
-        """C#용 ContextExtractor 인스턴스를 반환합니다."""
-        return ContextExtractor("csharp")
+        """Swift용 ContextExtractor 인스턴스를 반환합니다."""
+        return ContextExtractor("swift")
 
     def test_class_declaration(
         self,
@@ -33,41 +31,30 @@ class TestBasicFunctionExtraction:
         sample_file_path: Path,
     ) -> None:
         """클래스 선언부 추출 테스트."""
-        changed_ranges = [LineRange(32, 32)]  # SampleCalculator 클래스 선언부
+        changed_ranges = [LineRange(25, 25)]  # SampleCalculator 클래스 선언부
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        # using 문 검증
-        assert "using System;" in all_context
-        assert "using System.Collections.Generic;" in all_context
-        assert "using System.Linq;" in all_context
-        assert (
-            "public class SampleCalculator" in all_context
-            or "class SampleCalculator" in all_context
-        )
+        # import 문 검증
+        assert "import Foundation" in all_context
+        assert "class SampleCalculator" in all_context
 
     def test_constructor_method(
         self,
         extractor: ContextExtractor,
         sample_file_path: Path,
     ) -> None:
-        """생성자 메서드 추출 테스트."""
-        changed_ranges = [LineRange(46, 54)]  # 생성자
+        """init 메서드 추출 테스트."""
+        changed_ranges = [LineRange(34, 41)]  # init 메서드
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        # using 문 검증
-        assert "using System;" in all_context
-        assert "using System.Collections.Generic;" in all_context
-        assert "using System.Linq;" in all_context
-        # 부모 클래스 선언부 검증
-        assert "public class SampleCalculator" in all_context
-        assert (
-            "public SampleCalculator" in all_context
-            or "SampleCalculator(" in all_context
-        )
+        # import 문 검증
+        assert "import Foundation" in all_context
+        assert "class SampleCalculator" in all_context
+        assert "init(" in all_context
         assert "value" in all_context
         assert "history" in all_context
 
@@ -77,19 +64,15 @@ class TestBasicFunctionExtraction:
         sample_file_path: Path,
     ) -> None:
         """클래스 메서드 추출 테스트."""
-        changed_ranges = [LineRange(56, 90)]  # AddNumbers 메서드
+        changed_ranges = [LineRange(43, 72)]  # addNumbers 메서드
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        # using 문 검증
-        assert "using System;" in all_context
-        assert "using System.Collections.Generic;" in all_context
-        assert "using System.Linq;" in all_context
-        # 부모 클래스 선언부 검증
-        assert "public class SampleCalculator" in all_context
-        assert "AddNumbers" in all_context
-        assert "public" in all_context or "private" in all_context
+        # import 문 검증
+        assert "import Foundation" in all_context
+        assert "class SampleCalculator" in all_context
+        assert "func addNumbers" in all_context
 
     def test_complex_method(
         self,
@@ -97,18 +80,13 @@ class TestBasicFunctionExtraction:
         sample_file_path: Path,
     ) -> None:
         """복잡한 메서드 추출 테스트."""
-        changed_ranges = [LineRange(92, 149)]  # MultiplyAndFormat 메서드
+        changed_ranges = [LineRange(74, 126)]  # multiplyAndFormat 메서드
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        # using 문 검증
-        assert "using System;" in all_context
-        assert "using System.Collections.Generic;" in all_context
-        assert "using System.Linq;" in all_context
-        # 부모 클래스 선언부 검증
-        assert "public class SampleCalculator" in all_context
-        assert "MultiplyAndFormat" in all_context
+        assert "class SampleCalculator" in all_context
+        assert "func multiplyAndFormat" in all_context
 
     def test_nested_inner_function(
         self,
@@ -116,42 +94,44 @@ class TestBasicFunctionExtraction:
         sample_file_path: Path,
     ) -> None:
         """중첩 내부 함수 추출 테스트."""
-        changed_ranges = [LineRange(107, 114)]  # MultiplyRecursive 내부 함수
+        changed_ranges = [LineRange(86, 91)]  # multiplyRecursive 내부 함수
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
         # 부모 클래스 선언부 검증
-        assert "public class SampleCalculator" in all_context
-        # 부모 메서드 선언부 검증
-        assert "MultiplyAndFormat" in all_context
-        assert "MultiplyRecursive" in all_context
+        assert "class SampleCalculator" in all_context
+        assert "func multiplyAndFormat" in all_context
+        assert (
+            "func multiplyRecursive" in all_context
+            or "multiplyRecursive" in all_context
+        )
 
     def test_external_function(
         self,
         extractor: ContextExtractor,
         sample_file_path: Path,
     ) -> None:
-        """정적 메서드 추출 테스트."""
-        changed_ranges = [LineRange(179, 195)]  # HelperFunction
+        """외부 함수 추출 테스트."""
+        changed_ranges = [LineRange(148, 164)]  # helperFunction
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        assert "static" in all_context and "HelperFunction" in all_context
+        assert "func helperFunction" in all_context
 
     def test_factory_function(
         self,
         extractor: ContextExtractor,
         sample_file_path: Path,
     ) -> None:
-        """팩토리 메서드 추출 테스트."""
-        changed_ranges = [LineRange(204, 230)]  # AdvancedCalculatorFactory
+        """팩토리 함수 추출 테스트."""
+        changed_ranges = [LineRange(166, 188)]  # advancedCalculatorFactory
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        assert "AdvancedCalculatorFactory" in all_context
+        assert "func advancedCalculatorFactory" in all_context
 
     def test_method_declaration_only(
         self,
@@ -159,27 +139,26 @@ class TestBasicFunctionExtraction:
         sample_file_path: Path,
     ) -> None:
         """메서드 선언부만 추출 테스트."""
-        changed_ranges = [LineRange(56, 56)]  # AddNumbers 선언부만
+        changed_ranges = [LineRange(43, 43)]  # addNumbers 선언부만
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        # 부모 클래스 선언부 검증
-        assert "public class SampleCalculator" in all_context
-        assert "AddNumbers" in all_context
+        assert "class SampleCalculator" in all_context
+        assert "addNumbers" in all_context
 
     def test_external_function_declaration_only(
         self,
         extractor: ContextExtractor,
         sample_file_path: Path,
     ) -> None:
-        """정적 메서드 선언부만 추출 테스트."""
-        changed_ranges = [LineRange(179, 179)]  # HelperFunction 선언부만
+        """외부 함수 선언부만 추출 테스트."""
+        changed_ranges = [LineRange(148, 148)]  # helperFunction 선언부만
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        assert "HelperFunction" in all_context
+        assert "helperFunction" in all_context
 
 
 class TestModuleLevelElements:
@@ -190,15 +169,13 @@ class TestModuleLevelElements:
         """테스트용 샘플 파일 경로를 반환합니다."""
         return (
             Path(__file__).parent
-            / "language_samples"
-            / "csharp"
-            / "SampleCalculator.cs"
+            / "SampleCalculator.swift"
         )
 
     @pytest.fixture
     def extractor(self) -> ContextExtractor:
-        """C#용 ContextExtractor 인스턴스를 반환합니다."""
-        return ContextExtractor("csharp")
+        """Swift용 ContextExtractor 인스턴스를 반환합니다."""
+        return ContextExtractor("swift")
 
     def test_basic_constants(
         self,
@@ -206,26 +183,26 @@ class TestModuleLevelElements:
         sample_file_path: Path,
     ) -> None:
         """기본 상수들 추출 테스트."""
-        changed_ranges = [LineRange(10, 22)]  # Constants 클래스
+        changed_ranges = [LineRange(8, 10)]  # 상수 선언
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        assert "MAX_CALCULATION_STEPS" in all_context or "const" in all_context
-        assert "DEFAULT_PRECISION" in all_context
+        assert "let" in all_context
+        assert "100" in all_context or "2" in all_context
 
     def test_object_constant(
         self,
         extractor: ContextExtractor,
         sample_file_path: Path,
     ) -> None:
-        """객체 상수 추출 테스트."""
-        changed_ranges = [LineRange(16, 21)]  # CALCULATION_MODES 객체
+        """구조체 상수 추출 테스트."""
+        changed_ranges = [LineRange(12, 16)]  # CALCULATION_MODES 배열
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        assert "CALCULATION_MODES" in all_context
+        assert "CALCULATION_MODES" in all_context or "let" in all_context
 
     def test_module_bottom_constants(
         self,
@@ -233,7 +210,7 @@ class TestModuleLevelElements:
         sample_file_path: Path,
     ) -> None:
         """모듈 하단 상수들 추출 테스트."""
-        changed_ranges = [LineRange(234, 242)]  # ModuleConstants
+        changed_ranges = [LineRange(191, 195)]  # MODULE_VERSION
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
@@ -249,32 +226,30 @@ class TestMultiRangeExtraction:
         """테스트용 샘플 파일 경로를 반환합니다."""
         return (
             Path(__file__).parent
-            / "language_samples"
-            / "csharp"
-            / "SampleCalculator.cs"
+            / "SampleCalculator.swift"
         )
 
     @pytest.fixture
     def extractor(self) -> ContextExtractor:
-        """C#용 ContextExtractor 인스턴스를 반환합니다."""
-        return ContextExtractor("csharp")
+        """Swift용 ContextExtractor 인스턴스를 반환합니다."""
+        return ContextExtractor("swift")
 
     def test_three_cross_functions(
         self,
         extractor: ContextExtractor,
         sample_file_path: Path,
     ) -> None:
-        """3개 메서드에 걸친 영역 추출 테스트."""
+        """3개 함수에 걸친 영역 추출 테스트."""
         changed_ranges = [
-            LineRange(151, 230)
-        ]  # CalculateCircleArea ~ AdvancedCalculatorFactory
+            LineRange(128, 188)
+        ]  # calculateCircleArea ~ advancedCalculatorFactory
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
-        assert "CalculateCircleArea" in all_context
-        assert "HelperFunction" in all_context
-        assert "AdvancedCalculatorFactory" in all_context
+        assert "calculateCircleArea" in all_context
+        assert "helperFunction" in all_context
+        assert "advancedCalculatorFactory" in all_context
 
     def test_two_blocks_cross_methods(
         self,
@@ -282,13 +257,13 @@ class TestMultiRangeExtraction:
         sample_file_path: Path,
     ) -> None:
         """2개 블록에 걸친 메서드 추출 테스트."""
-        changed_ranges = [LineRange(56, 64), LineRange(179, 185)]
+        changed_ranges = [LineRange(43, 51), LineRange(148, 153)]
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
-        assert len(contexts) >= 1
+        assert len(contexts) >= 2
         all_context = "\n".join(contexts)
-        assert "AddNumbers" in all_context
-        assert "HelperFunction" in all_context
+        assert "addNumbers" in all_context
+        assert "helperFunction" in all_context
 
     def test_non_contiguous_ranges(
         self,
@@ -297,15 +272,15 @@ class TestMultiRangeExtraction:
     ) -> None:
         """비연속적인 여러 범위 추출 테스트."""
         changed_ranges = [
-            LineRange(12, 14),  # 파일 상수들
-            LineRange(158, 161),  # ValidateRadius 내부 함수
-            LineRange(236, 238),  # 모듈 레벨 상수들
+            LineRange(8, 10),  # 파일 상수들
+            LineRange(134, 136),  # validateRadius 내부 함수
+            LineRange(191, 195),  # 모듈 레벨 상수들
         ]
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
-        assert len(contexts) >= 1
+        assert len(contexts) >= 2
         all_context = "\n".join(contexts)
-        assert "MAX_CALCULATION_STEPS" in all_context or "const" in all_context
+        assert "MAX_CALCULATION_STEPS" in all_context or "let" in all_context
         assert "MODULE_VERSION" in all_context
 
 
@@ -317,15 +292,13 @@ class TestComplexScenarios:
         """테스트용 샘플 파일 경로를 반환합니다."""
         return (
             Path(__file__).parent
-            / "language_samples"
-            / "csharp"
-            / "SampleCalculator.cs"
+            / "SampleCalculator.swift"
         )
 
     @pytest.fixture
     def extractor(self) -> ContextExtractor:
-        """C#용 ContextExtractor 인스턴스를 반환합니다."""
-        return ContextExtractor("csharp")
+        """Swift용 ContextExtractor 인스턴스를 반환합니다."""
+        return ContextExtractor("swift")
 
     def test_entire_class_extraction(
         self,
@@ -333,14 +306,14 @@ class TestComplexScenarios:
         sample_file_path: Path,
     ) -> None:
         """전체 클래스 추출 테스트."""
-        changed_ranges = [LineRange(32, 171)]  # SampleCalculator 전체 클래스
+        changed_ranges = [LineRange(25, 146)]  # SampleCalculator 전체 클래스
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
         all_context = "\n".join(contexts)
         assert "class SampleCalculator" in all_context
-        assert "SampleCalculator(" in all_context
-        assert "AddNumbers" in all_context
+        assert "init(" in all_context
+        assert "addNumbers" in all_context
 
     def test_class_and_module_constants(
         self,
@@ -348,7 +321,7 @@ class TestComplexScenarios:
         sample_file_path: Path,
     ) -> None:
         """클래스와 모듈 상수 동시 추출 테스트."""
-        changed_ranges = [LineRange(10, 242)]  # 상수부터 모듈 끝까지
+        changed_ranges = [LineRange(8, 195)]  # 상수부터 모듈 끝까지
         contexts = extractor.extract_contexts(sample_file_path, changed_ranges)
 
         assert len(contexts) >= 1
@@ -365,15 +338,13 @@ class TestEdgeCases:
         """테스트용 샘플 파일 경로를 반환합니다."""
         return (
             Path(__file__).parent
-            / "language_samples"
-            / "csharp"
-            / "SampleCalculator.cs"
+            / "SampleCalculator.swift"
         )
 
     @pytest.fixture
     def extractor(self) -> ContextExtractor:
-        """C#용 ContextExtractor 인스턴스를 반환합니다."""
-        return ContextExtractor("csharp")
+        """Swift용 ContextExtractor 인스턴스를 반환합니다."""
+        return ContextExtractor("swift")
 
     def test_invalid_line_ranges(
         self,
